+++ crates/floresta-rpc/src/rpc_interfaces.rs
@@ -171,6 +171,7 @@ pub trait NetworkRpc {
     fn ping(&self) -> Result<bool, Self::Error>;
 }
 
+#[floresta_proc_macro::maybe_async]
 pub trait RawTransactionRpc {
     type Error: Display + Debug;
